+++ R/scenario.R
@@ -43,7 +43,8 @@ methods::setMethod(
   function(fit, limits = NULL, reuse_limits = FALSE, copy_model = FALSE) {
     # Check that arguments are valid
     assertthat::assert_that(!missing(fit) || inherits(fit,'DistributionModel'),
-                            inherits(limits, 'SpatRaster') || inherits(limits, 'sf') || inherits(limits, 'Spatial') || is.null(limits),
+                            inherits(limits, 'SpatRaster') || inherits(limits, 'sf') || inherits(limits, 'Spatial') ||
+                              is.null(limits) || inherits(limits, "stars"),
                             is.logical(reuse_limits),
                             is.logical(copy_model),
                             msg = 'No trained model supplied!')
@@ -63,15 +64,16 @@ methods::setMethod(
       if(inherits(limits,'SpatRaster')){
         # Remove 0 from ratified raster assuming this is no-data
         limits[limits == 0] <- NA
-        limits <- sf::st_as_sf( terra::as.polygons(limits, trunc = TRUE, dissolve = TRUE) )
+        limits <- terra_to_sf(limits)
       }
       # Ensure that limits has the same projection as background
       if(sf::st_crs(limits) != sf::st_crs(fit$model$background)) limits <- sf::st_transform(limits, fit$model$background)
       # Ensure that limits is intersecting the background
       if(suppressMessages(length( sf::st_intersects(limits, fit$model$background)))==0) { limits <- NULL; warning('Provided limits do not intersect the background!') }
 
-      # Get fir column and rename
-      limits <- limits[,1]; names(limits) <- c('limit','geometry')
+      # Rename geometry just to be sure
+      limits <- rename_geometry(limits, "geometry")
+      names(limits)[1] <- "limit" # The first entry is the layer name
     }
 
     # Also check if limits are to be reused if found